pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genai-explainer"
version = "0.1.0"
description = "Offline paper-to-explainer: PDF -> outline -> narrated segments -> slideshow MP4"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "python-multipart>=0.0.9",
  "celery>=5.3",
  "redis>=5.0",
  "pypdf>=4.0",
  "rank-bm25>=0.2.2",
  "pillow>=10.0",
  "httpx>=0.27",
  "duckduckgo-search>=6.3.0",
  "playwright>=1.41.2",
  "psycopg2-binary>=2.9.9",
  "google-cloud-texttospeech>=2.16.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "core*", "db*"]
exclude = ["data*", "infra*", "scripts*", "tests*"]
